% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bleachCorrect.R
\name{bleachCorrect}
\alias{bleachCorrect}
\title{bleachCorrect}
\usage{
bleachCorrect(data.object, rec.frames = NA, bleach.begin = 1,
  bleach.leaveout = NA, weight.BS = 8, weight.AS = 1,
  polcol = "#cccccc", mfrow = c(5, 4), fixedA = NA, fixedB = NA,
  fixedC = NA, ref = "MOL", estimate.B = F, bleach.begin.ref = NA,
  bleach.leaveout.ref = NA, weight.BS.ref = NA, weight.AS.ref = NA,
  diffParms4Ref = F, plot = T, ylim = NULL)
}
\arguments{
\item{data.object}{object containing the data}

\item{rec.frames}{number of frames recorded}

\item{bleach.begin}{number of the startframe for the fitting (defaults to 1)}

\item{bleach.leaveout}{numerical vector of frames to leave out for fitting (where the stimulus comes) (e.g.: c(27:65))}

\item{weight.BS}{how to weight the part before the stimulus}

\item{weight.AS}{how to weight the part after the stimulus}

\item{polcol}{color of the weight polygon}

\item{mfrow}{}

\item{fixedA}{fixed value for tparameter A}

\item{fixedB}{fixed value for tparameter B}

\item{fixedC}{fixed value for tparameter C}

\item{ref}{control to calculate the bleach correction on}

\item{estimate.B}{estimate parameter B from fits to the control?}

\item{bleach.begin.ref}{number of the startframe for the fitting (specific for fits to control)}

\item{bleach.leaveout.ref}{umerical vector of frames to leave out for fitting (where the stimulus comes) (e.g.: c(27:65))}

\item{weight.BS.ref}{how to weight the part before the stimulus (specific for fits to control)}

\item{weight.AS.ref}{how to weight the part after the stimulus (specific for fits to control)}

\item{diffParms4Ref}{define different fitting parameters for the reference (MOL)}

\item{plot}{plot?}

\item{ylim}{y limits for plots}
}
\description{
correct bleaching in gloDatamix data
}
\author{
Daniel Münch <daniel@muench.bio>
}

