% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctvs.R
\name{ctvs}
\alias{ctvs}
\title{ctvs}
\usage{
ctvs(data.object, ctv.type, rec.frames = NA, bg.firstframe = NA,
  bg.lastframe = NA, sig.firstframe = NA, sig.lastframe = NA,
  wa = c(0.2, 0.7), plot = F, report = F, ylim = NA, mfrow = c(4, 3),
  suffix = "", meanTraces = F, fp = c(29, 37, 41, 49))
}
\arguments{
\item{data.object}{input data (gloDatamix format)}

\item{ctv.type}{type of CTV to calculate}

\item{rec.frames}{number of frames recorded, will be autodetected if empty}

\item{bg.firstframe}{for CTVs with background subtraction this is the position of the first bg frame}

\item{bg.lastframe}{2nd frame for background (end of window)}

\item{sig.firstframe}{signal onset}

\item{sig.lastframe}{signal offset}

\item{wa}{calc "width @" xx*extremum, needs 2 arguments, e.g. width when 0.2 of peakmax is reached to when peak falls to .9 of peakmax}

\item{plot}{perform a barplot in the end?}

\item{report}{write CTV report file (gives single traces with signal, bg ... marked and values in legend)}

\item{ylim}{ylim for report plots, given as 2 value vector, e.g. c(-1,5) plots from -1 to}

\item{mfrow}{numer of columns and rows to plot per page}

\item{suffix}{optionally, ad describing suffix to filenames, name of data.object is used when empty}

\item{meanTraces}{does data come from meanTraces.R?}

\item{fp}{numerical vector of length 4 giving positions for ctv.w for putative peak on and offset}
}
\description{
collection of 'CurveToValue' functions
}
\author{
Daniel Münch <daniel@muench.bio>
}

