% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findDuplicates.R
\name{findDuplicates}
\alias{findDuplicates}
\title{find duplicated odor names}
\usage{
findDuplicates(data.object, case = "upper", molsingle = F,
  usualsuspects = T, conc0 = T, cut.odors = F, conc.from.bc = F,
  conc.from.odor = F)
}
\arguments{
\item{data.object}{input data in gloDatamix format}

\item{case}{transform all to "upper", "lower" or "common" case, (common not implemented yet); or "test" only}

\item{molsingle}{find misspelled MOL recordings (only for single recordings, DONT USE WITH TWINPAL), greps for "mol" and replaces all hits with "mol"}

\item{usualsuspects}{correct for some ususal made misspellings like BJOT BIOT, ALOT AIOT}

\item{conc0}{set NOConc for "AIR" and "MOL" to 0}

\item{cut.odors}{remove everything beyond 4 letters from the odor names}

\item{conc.from.bc}{read concentration from barcode -- !!handle with care!!}

\item{conc.from.odor}{read concentration from odorfield, assuming is is the last 2 digits -- !!!handle with even more care!!!}
}
\description{
checks 4 letter odor codes for duplications
}
\author{
Daniel Münch <daniel@muench.bio>
}

