% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDuplicates_mix.R
\name{FindDuplicates_mix}
\alias{FindDuplicates_mix}
\title{Find Duplicates Mix}
\usage{
FindDuplicates_mix(data.object, case = "upper", molsingle = F, conc0 = T)
}
\arguments{
\item{data.object}{input data frame}

\item{conc0}{}
}
\description{
looks for duplicates in binary mixture codes
}

