% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloQplot.R
\name{gloQplot}
\alias{gloQplot}
\title{gloQplot}
\usage{
gloQplot(data.object, ylim = NULL, mean = F, median = F, legend = F,
  ...)
}
\arguments{
\item{data.object}{gloDatamix input data.frame}

\item{ylim}{plotting limit}

\item{mean}{plot mean trace?}

\item{median}{plot median trace?}

\item{legend}{plot a legend?}

\item{...}{pass additional parameters to the plot() function}
}
\value{
plot
}
\description{
quick and dirty plots from gloDatamix data
}
\author{
Daniel Münch <daniel@muench.bio>
}

