% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCTV.R
\name{meanCTV}
\alias{meanCTV}
\title{mean CTV}
\usage{
meanCTV(data.object, test = F, nullref = "MOL", plev = 0.05,
  ctvcolumnname = "ctv", shift = F)
}
\arguments{
\item{data.object}{gloDatamix data.frame derived from ctvs.R}

\item{test}{logical, perform t.test() against control?}

\item{nullref}{name of the control}

\item{plev}{plevel to test against}

\item{ctvcolumnname}{name of the ctv containing column}

\item{shift}{subtract control responses from values?}
}
\description{
Function to calculate mean, median... from ctvs, needs output from ctvs.R as input (gloDatamix with ctv column)
}
\author{
Daniel Münch <daniel@muench.bio>
}

