% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanTracesGlom.R
\name{meanTraces}
\alias{meanTraces}
\title{meanTraces}
\usage{
meanTraces(data.object, rec.frames = NA, st.shift = NA, ylim.lower = NA,
  ylim.upper = NA, add = "single", method = "mean", pdf = T,
  mfrow = c(4, 3), receptor = NA, return = TRUE)
}
\arguments{
\item{data.object}{gloDatamix data.frame}

\item{rec.frames}{number of frames measured}

\item{st.shift}{number of frames to shift the stimulus bars}

\item{ylim.lower}{limits of the y axis, if nothing is specified the max/min of the dataset will be used; it is also okay to only specify one limit, then only the other one will be calculated}

\item{ylim.upper}{limits of the y axis, if nothing is specified the max/min of the dataset will be used; it is also okay to only specify one limit, then only the other one will be calculated}

\item{add}{also plot single recordings ("single"), standard deviation ("sd") or standard error of mean ("sem"); default is "single"}

\item{method}{use "median" or "mean", default is "mean"}

\item{pdf}{plot into a PDF?}

\item{mfrow}{nr of rows/columns to plot}

\item{receptor}{adds a column containing the receptor name provided here}

\item{return}{return the results in the end or jus produce a pdf?}
}
\value{
list and PDF
}
\description{
calculate and plot mean traces from gloDatamix data
}
\author{
Daniel Münch <daniel@muench.bio>
}

