% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeToReference.R
\name{normalizeToReference}
\alias{normalizeToReference}
\title{normalizeToReference}
\usage{
normalizeToReference(data.object, rec.frames = NA, reference.odor, ctv.type,
  bg.firstframe = NA, bg.lastframe = NA, sig.firstframe, sig.lastframe,
  pdf = T, mode = "mean", rescale.data = F, ctvreport = F, glom = NA,
  suffix = "", plot = T)
}
\arguments{
\item{data.object}{gloDatamix data.frame}

\item{rec.frames}{number of recorded frames}

\item{reference.odor}{your positive reference}

\item{ctv.type}{the type of ctv used for normalization}

\item{bg.firstframe}{first frame of background (passed to ctvs())}

\item{bg.lastframe}{last frame of background (passed to ctvs())}

\item{sig.firstframe}{first frame of the signal (passed to ctvs())}

\item{sig.lastframe}{last frame of the signal (passed to ctvs())}

\item{pdf}{produce a PDF of the fits?}

\item{mode}{one of 'mean', 'linFit', 'linFitEachGlom'}

\item{rescale.data}{if FALSE: reference set to 1, if set to TRUE: values are multiplied with first reference value afterwards}

\item{ctvreport}{create a ctv report PDF?}

\item{glom}{if specified the data will be normalized on this single glomerulus}

\item{suffix}{suffix for the filename}

\item{plot}{plot the fits?}
}
\description{
function that fits a curve over reference odors and normalizes to it
}
\details{
Scheme: 1. subset of reference
2. ctv of reference \cr
3. fit curve over ctv/time \cr
A-4. divide every timepoint of every measurement through corresponding point on fitted curve -> normalized curves \cr
A-5. do ctv for measurements -> normalized ctvs \cr
 \cr
B-4. do ctv for measurements \cr
B-5. divide every ctv through corresponding point on fitted curve \cr
}
\author{
Daniel Münch <daniel@muench.bio>
}

