% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overviewGlodatamix.R
\name{overviewGlodatamix}
\alias{overviewGlodatamix}
\title{overview gloDatamix}
\usage{
overviewGlodatamix(data.object, rec.frames = NA, ref = NA, mol = NA,
  air = NA, ylim_lower = NA, ylim_upper = NA, mfrow = c(4, 3),
  glomwise = F, odorandglomwise = F, stimulusframes = NA,
  polcol = "#efefef", suffix = "")
}
\arguments{
\item{data.object}{input data.fram (gloDatamix)}

\item{rec.frames}{number of recorded frames, read from gloDatamix file if not specified}

\item{ref}{name of your reference, if specified, all references will be plottet in 1st plot}

\item{mol}{name of your mineral oil control, if specified, all mols will be plottet in 1st plot}

\item{air}{name of your room-air conrol, if specified, all airs will be plottet in 1st plot}

\item{ylim_lower}{lower y axis limit, if empty minimum value is taken}

\item{ylim_upper}{upper y axis limit, if empty maximum value is taken}

\item{mfrow}{number of rows and columns per page}

\item{glomwise}{plot glomerulus wise}

\item{odorandglomwise}{lot odor and glomerulus wise}

\item{stimulusframes}{only needed when stimulus is different from what is in the gloDatamix file, used to plot a polygon. has to be defined as one vector stim. start and stop frames ("c(20,24,30,36)")}

\item{polcol}{color of the stimulus polygon}

\item{suffix}{suffic for filenames}
}
\description{
prints an pdf overview per animal with additionally all mol, air and reference measurements in one plot
}
\author{
Daniel Münch <daniel@muench.bio>
}

