% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGlodatamix.R
\name{read.gdm}
\alias{read.gdm}
\title{read.gdm}
\usage{
read.gdm(path, header = TRUE, input = "gloDatamix")
}
\arguments{
\item{path}{character, location of the *.gloDatamix files}

\item{header}{logical, gloDataMix files with or without header?}

\item{input}{character, 'gloDatamix' or 'csv'}
}
\value{
concatenates all gloDatamix files from a location in one data.frame
}
\description{
read.gdm
}
\examples{
data <- read.gdm("./80109-final_all/", header=F)
}
\author{
Daniel Münch <daniel@muench.bio>
}

